Changed lowestCost data type to int in GetSortedProducts_Valid_Cost_Selected_Should_Return_Cities_Sorted_By_Cost unit test in Productlist.Razor.Tests.cs

+++ UnitTests/Components/ProductList.razor.Tests.cs
@@ -741,7 +741,7 @@ namespace UnitTests.Components
             int highestCost = productModels.Max(x => x.Cost);
 
             // The lowest cost in the database
-            double lowestCost = productModels.Min(x => x.Cost);
+            int lowestCost = productModels.Min(x => x.Cost);
 
             // Find the productModels with the highest cost
             var mostExpensiveCities = productModels.Where(x => x.Cost == highestCost);
